[PATCH] epca and specialix warning fixes
Event is just used for internal flags and with set_bit for atomicity. This
kills the warning the obvious way

+++ drivers/char/specialix_io8.h
@@ -110,7 +110,7 @@ struct specialix_port {
 	struct tty_struct 	* tty;
 	int			count;
 	int			blocked_open;
-	int			event;
+	ulong			event;
 	int			timeout;
 	int			close_delay;
 	long			session;
